+++ doc/source/doxygen-docs/changelog.md
@@ -6,6 +6,8 @@
     - Both mrpt::containers::CDynamicGrid and mrpt::containers::CDynamicGrid3D are now compatible with range-based for loops, and also have a data() method.
   - \ref mrpt_core_grp
     - Added mrpt::LockHelper::unlock()
+  - \ref mrpt_gui_grp
+    - New method mrpt::gui::CGlCanvasBase::CamaraParams::FromCamera()
   - \ref mrpt_math_grp
     - Added missing method for consistent API across pose classes: mrpt::math::TPose3D::operator+()
 - BUG FIXES: